Wire up Loading dots for async calls
Whenever an async call is in progress, we should display the loading
dots component so they know something is happening. The logic behind
this is counting the number of ajax calls in progress. I don't
particularly like this solution because it requires the developer to
remember to dispatch the action whenever a new ajax based function
is implemented.

At the moment, this is not working because the internal calls made by
redux mess up the count. I will investigate this further when I get
a better understanding of how react and redux work :) For now, let the
learning process continue :dancers: